renamed schemas

+++ src/private/words/words.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { MWord, WordDocument } from 'src/private/words/words.schema';
+import { Word, WordDocument } from 'src/private/words/words.schema';
 import {
   stringUtilExistsMessage as existMessage,
   stringUtilsNotExistsMessage as notExistMessage,
@@ -12,20 +12,9 @@ import { CreateWordDto, UpdateWordDto } from 'src/private/words/words.dto';
 @Injectable()
 export class WordsService {
   constructor(
-    @InjectModel(MWord.name, 'dictDB')
+    @InjectModel(Word.name, 'dictDB')
     private wordModel: Model<WordDocument>,
   ) {}
-
-  async createWord(newWord: CreateWordDto) {
-    const wordExists = await this.wordModel.findOne({
-      value: newWord.value,
-    });
-    if (wordExists) {
-      throw new Error(existMessage('Word', newWord.value));
-    }
-    const newWordDoc = await new this.wordModel(newWord).save();
-    return { message: 'Word created', word: newWordDoc };
-  }
 
   async getWord(word: string) {
     const wordDoc = await this.wordModel.findOne({
@@ -37,6 +26,15 @@ export class WordsService {
     }
 
     return wordDoc;
+  }
+
+  async createWord(newWord: CreateWordDto) {
+    const wordExists = this.getWord(newWord.value);
+    if (!wordExists) {
+      throw new Error(existMessage('Word', newWord.value));
+    }
+    const newWordDoc = await new this.wordModel(newWord).save();
+    return { message: 'Word created', word: newWordDoc };
   }
 
   async updateWord(update: UpdateWordDto, word: string) {
@@ -51,6 +49,9 @@ export class WordsService {
 
   async deleteWord(word: string) {
     const wordDoc = this.getWord(word);
+    if (!wordDoc) {
+      throw new Error(notExistMessage('Word', word));
+    }
     await this.wordModel.findByIdAndDelete(wordDoc);
     return { message: wasDeletedMessage('Word', word) };
   }
